Implemented total price for 3rd tab.

+++ CUWebinars.Web/ViewModel/OrderHasAdditionalLocationsViewModel.cs
@@ -7,6 +7,6 @@ namespace CUWebinars.Web.ViewModel
     {
         public IEnumerable<AdditionalLocation> AdditionalLocations { get; set; }
         public string Addresses { get; set; }
-        public int OptionsCost { get; set; }
+        public decimal OptionsCost { get; set; }
     }
 }
